[⭐] Added SHA512 and improved efficiency with rayon
significant improvements to the hasher program. We’ve added SHA512 to the list of hash algorithms, alongside the existing SHA256 and MD5. To boost efficiency, we’ve also incorporated the use of a thread pool via the rayon crate for concurrent computation of hashes. This update enhances the program’s performance and capabilities.